Issues Fixes. - Cart Screen Tax Total and Tax list mis match  - Dine in order table screen total pay amount 0.01 difference from cart screen.  - Guest total and pay total amount differnece  - Dine in Checkout Screen Total Tax and List tax 0.01 difference.

+++ app/src/main/java/com/android/pos/utils/MethodUtils.kt
@@ -602,11 +602,11 @@ class MethodUtils {
                 if (tmp.length > tmpIndex + 3) {
 
 
-                    return String.format("%.2f", value).toBigDecimal().toDouble()
+                    return String.format("%.2f", value).toDouble()
 
                 } else {
 
-                    return String.format("%.2f", value).toBigDecimal().toDouble()
+                    return String.format("%.2f", value).toDouble()
 
                 }
             } catch (e: java.lang.Exception) {
